Organise data into channels
We recieve the data similar to the format of SerialDataPoint, but the hot path is drawing the GUI, which wants it in a more channel-oriented format.
This tidies up the code a little. Don't need `checked` and `colours` and `data` floating around separately.
This does lead to a little bit of data duplication though, as the timestamp is included in each channel's datapoints, rather than being stored centrally.
Also use a separate thread to save the CSV and clone the values in there, rather than in the GUI thread.